requirements.txt
PySide6==6.11.0
pandas==3.0.2
openpyxl==3.1.5
python-dotenv==1.2.2
pyinstaller==6.19.0
pyodbc==5.3.0
pywin32==311
